refactor: move components, theme and hooks to site
- remove packages/ui
- remove packages/stitches-config
