--- packages/stitches-config/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@junat/stitches",
-  "version": "0.0.1",
-  "main": "dist/stitches.config.js",
-  "types": "dist/stitches.config.d.ts",
-  "private": true,
-  "scripts": {
-    "build": "shx rm -rf dist && tsc --build"
-  },
-  "dependencies": {
-    "@stitches/react": "^1.2.8",
-    "react": "^18.1.0"
-  },
-  "devDependencies": {
-    "@junat/prettier": "workspace:*",
-    "@types/react": "^18.0.9",
-    "shx": "^0.3.4",
-    "typescript": "^4.6.4"
-  },
-  "prettier": "@junat/prettier"
-}
