--- packages/ui/.storybook/main.js
@@ -1,15 +1,0 @@
-module.exports = {
-  stories: [
-    '../src/stories/**/*.stories.mdx',
-    '../src/stories/**/*.stories.@(js|jsx|ts|tsx)'
-  ],
-  addons: [
-    '@storybook/addon-links',
-    '@storybook/addon-essentials',
-    '@storybook/addon-interactions'
-  ],
-  framework: '@storybook/react',
-  core: {
-    builder: 'webpack5'
-  }
-}
