--- packages/ui/.storybook/preview.js
@@ -1,15 +1,0 @@
-import './reset.css'
-
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' }
-}
-
-export const decorators = [
-  Story => {
-    return (
-      <div style={{ maxWidth: '550px', margin: 'auto' }}>
-        <Story />
-      </div>
-    )
-  }
-]
