--- packages/ui/.storybook/reset.css
@@ -1,44 +1,0 @@
-/* Preferred box-sizing value */
-*,
-*::before,
-*::after {
-  box-sizing: border-box;
-}
-
-/*
-      Remove list styles (bullets/numbers)
-      in case you use it with normalize.css
-    */
-ol,
-ul {
-  list-style: none;
-}
-
-/* For images to not be able to exceed their container */
-img {
-  max-width: 100%;
-}
-
-/* Removes spacing between cells in tables */
-table {
-  border-collapse: collapse;
-}
-
-/* Revert the 'white-space' property for textarea elements on Safari */
-textarea {
-  white-space: revert;
-}
-
-.docs-story,
-html {
-  background-color: #fbfbfb;
-  all: unset;
-  display: revert;
-  color-scheme: light dark;
-}
-@media (prefers-color-scheme: dark) {
-  .docs-story,
-  html {
-    background-color: var(--junat-colors-secondary900);
-  }
-}
