--- packages/ui/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "@junat/ui",
-  "version": "0.0.1",
-  "main": "index.js",
-  "types": "index.d.ts",
-  "license": "MIT",
-  "sideEffects": false,
-  "scripts": {
-    "build": "shx rm -rf dist && tsc --build && pnpm copy",
-    "copy": "shx cp package.json dist && json -I -f dist/package.json -e \"this.devDependencies=undefined; this.scripts=undefined; this.prettier=undefined;\" && pnpm i -r",
-    "test": "vitest run",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
-  },
-  "publishConfig": {
-    "directory": "dist"
-  },
-  "dependencies": {
-    "@junat/digitraffic": "workspace:*",
-    "@junat/stitches": "workspace:*",
-    "framer-motion": "^6.3.3"
-  },
-  "peerDependencies": {
-    "react": ">=16.8",
-    "react-dom": ">=16.8"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.17.12",
-    "@junat/prettier": "workspace:*",
-    "@storybook/addon-actions": "^6.5.7",
-    "@storybook/addon-essentials": "^6.5.7",
-    "@storybook/addon-interactions": "^6.5.7",
-    "@storybook/addon-links": "^6.5.7",
-    "@storybook/builder-webpack5": "^6.5.7",
-    "@storybook/manager-webpack5": "^6.5.7",
-    "@storybook/react": "^6.5.7",
-    "@storybook/testing-library": "^0.0.11",
-    "@types/react": "^18.0.9",
-    "babel-loader": "^8.2.5",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "require-from-string": "^2.0.2",
-    "rimraf": "^3.0.2",
-    "shx": "^0.3.4",
-    "typescript": "^4.7.2",
-    "vitest": "^0.18.0",
-    "webpack": "^5"
-  },
-  "prettier": "@junat/prettier"
-}
