--- packages/ui/src/index.ts
@@ -1,4 +1,0 @@
-export * from './components/SingleTimetable'
-export * from './components/SingleTimetableRow'
-export * from './components/Timetable'
-export * from './components/TimetableRow'
