+++ site/src/config/theme/colors/slate_gray.ts
