+++ site/src/styles/global.ts
@@ -1,4 +1,4 @@
-import { globalCss } from '@junat/stitches'
+import { globalCss } from '@config/theme'
 
 const SYSTEM_FONTS =
   "-apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif" as const
