+++ site/src/styles/reset.ts
@@ -1,4 +1,4 @@
-import { globalCss } from '@junat/stitches'
+import { globalCss } from '@config/theme'
 
 export const reset = globalCss({
   '*:where(:not(iframe, canvas, img, svg, video):not(svg *))': {
